Arbitrary image usage, scales with number of tiles desired. Now, just needs interactivity.

+++ src/components.rs
@@ -4,5 +4,5 @@ use amethyst::{
 
 #[derive(Component)]
 pub struct Tile {
-    pub index: usize,
+    pub index: u32,
 }
